+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.7+2
+
+* Updates lint checks to ignore NewerVersionAvailable.
+
 ## 0.6.7+1
 
 * Updates README to remove references to `maxVideoDuration`, as it was never